Refactor conditional logic by putting in separate module
